+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.3.7] - 2020-01-30
+## Changed
+- Bulk revert is now supported.
+- Collab is blocked in play mode.
+## Fixed
+- Fixed services window's links to open Collab.
+
 ## [1.3.6] - 2020-01-21
 ### Fixed
 - Fixed compile errors when removing the NUnit package by removing unnecessary references.